Move inventory functions from Manager to Inventory
