+++ src/Inventory.js
@@ -7,6 +7,7 @@
 
         define( [
             'jquery',
+            './Util',
             './Config'
         ], factory );
 
@@ -14,6 +15,7 @@
 
         module.exports = factory(
             require( 'jquery' ),
+            require( './Util' ),
             require( './Config' )
         );
 
@@ -23,18 +25,158 @@
 
         root.AdManager.Inventory = factory(
             root.jQuery,
-            root.AdManager.Conifg
+            root.AdManager.Util,
+            root.AdManager.Config
         );
 
     }
 
-} ( this, function ( $, Config ) {
+} ( this, function ( $, Util, Config ) {
 
     var name = 'Inventory',
         debugEnabled = true,
         debug;
 
     //////////////////////////////////////////////////////////////////////////////////////
+
+    /**
+     * Get sanitized inventory.
+     *
+     * @return {Object}
+     */
+    function getInventory() {
+
+        return getAvailableSizes( inventoryCleanTypes( Config.get( 'inventory' ) ) );
+
+    }
+
+    /**
+     * Add default unit type if not set
+     *
+     * @param array inventory
+     * @return array inventory
+     */
+    function inventoryCleanTypes( inventory ) {
+
+        for ( var i = 0; i < inventory.length; i++ ) {
+
+            if ( typeof inventory[ i ].type !== 'undefined' ) {
+                continue;
+            }
+
+            inventory[ i ].type = 'default';
+
+        }
+
+        return inventory;
+
+    }
+
+    /**
+     * Remove sizes from inventory that will not display properly
+     *
+     * @param array inventory
+     * @return array inventory
+     */
+    function getAvailableSizes( inventory ) {
+
+        var width = ( window.innerWidth > 0 ) ? window.innerWidth : screen.width;
+
+        if ( width > 1024 ) {
+            return inventory;
+        }
+
+        if ( width >= 768 && width <= 1024 ) {
+
+            var max = 980;
+
+            for ( var i = 0; i < inventory.length; i++ ) {
+                var sizesToRemove = [];
+                for ( var j = 0; j < inventory[ i ].sizes.length; j++ ) {
+                    if ( inventory[ i ].sizes[ j ][0] > max ) {
+                        sizesToRemove.push( inventory[ i ].sizes[ j ] );
+                    }
+                }
+                inventory[ i ].sizes = Util.difference( inventory[ i ].sizes, sizesToRemove );
+            }
+        }
+
+        return inventory;
+
+    }
+
+    /**
+     * removeDefinedSlot
+     *
+     * @param string name
+     * @return object definedSlot
+     */
+    function removeDefinedSlot( name ) {
+
+        $.each( definedSlots, function ( index, slot ) {
+            var unitName = slot.getAdUnitPath().replace( '/' + account + '/', '' );
+            if ( unitName === name ) {
+                definedSlots = Util.removeByKey( definedSlots, index );
+            }
+        } );
+
+    }
+
+    function getDynamicInventory() {
+
+        var dynamicItems = [],
+            type = Config.get( 'clientType' ),
+            localContext;
+
+        $.each( inventory, function ( index, position ) {
+            if ( ( typeof position.dynamic !== 'undefined' ) && ( position.dynamic === true ) ) {
+                if ( ! type || type === position.type ) {
+                    dynamicItems.push( position );
+                    localContext = position.localContext;
+                }
+            }
+        } );
+
+        return {
+            dynamicItems: dynamicItems,
+            localContext: localContext
+        };
+
+    }
+
+    /**
+     * Get Ad Unit Info
+     *
+     * @param string unit
+     * @return object
+     */
+    function getAdInfo( unit ) {
+
+        var returnObject = {};
+
+        for ( var i = 0; i < inventory.length; i++ ) {
+            if ( inventory[ i ].id !== unit && inventory[ i ].slot !== unit ) {
+                continue;
+            }
+
+            // build return object
+            returnObject = inventory[ i ];
+
+            // determine the object's idName
+            if ( typeof returnObject.useIterator !== 'undefined' && ! returnObject.useIterator ) {
+                // don't use the iterator
+                returnObject.idName = returnObject.id;
+            } else {
+                // use the iterator
+                returnObject.idName = returnObject.id + '_' + returnObject.iteration;
+            }
+
+            return returnObject;
+        }
+
+        return returnObject;
+
+    }
 
     /**
      * Get Shortest Possible Size for Unit
@@ -127,10 +269,13 @@
     //////////////////////////////////////////////////////////////////////////////////////
 
     return {
-        shortestAvailable: shortestAvailable,
-        tallestAvailable:  tallestAvailable,
-        limitUnitHeight:   limitUnitHeight,
-        getUnitType:       getUnitType
+        getInventory:        getInventory,
+        getAdInfo:           getAdInfo,
+        getDynamicInventory: getDynamicInventory,
+        shortestAvailable:   shortestAvailable,
+        tallestAvailable:    tallestAvailable,
+        limitUnitHeight:     limitUnitHeight,
+        getUnitType:         getUnitType
     };
 
 } ) );
